+++ webpack.test.config.js
@@ -3,8 +3,13 @@ const webpackCommonConfig = require('./webpack.common.config.js');
 
 function config(opts) {
 	return Object.assign({}, webpackCommonConfig(opts), {
-		mode: "test"
+		mode: "development",
+		devtool: "inline-source-map"
 	});
 }
 
-module.exports = config;
+module.exports = config({
+	isTest: true,
+	isDevelopment: false,
+	isProduction: false
+});